requirements.txt
Pillow>=9.0.0
pyautogui>=0.9.50
